Make vk_usec_now inlined
